+++ Changes.md
@@ -8,6 +8,7 @@ you spot any mistakes.
 
 * Clone connection config for new pool connections
 * Default `connectTimeout` to 2 minutes
+* Throw on unknown SSL profile name #817
 
 ## v2.2.0 (2014-04-27)